Address comments
- Moved config functions in config.go
- Moved pprof stuff in pprof.go
- renamed NewVX to New.
- API version point to a constant now.
