+++ mgmt/rest/v2/api.go
@@ -26,13 +26,14 @@ type V2 struct {
 	killChan chan struct{}
 }
 
-func NewV2(wg *sync.WaitGroup, killChan chan struct{}, protocol string) *V2 {
+func New(wg *sync.WaitGroup, killChan chan struct{}, protocol string) *V2 {
 	protocolPrefix = protocol
 	return &V2{wg: wg, killChan: killChan}
 }
 
 func (s *V2) GetRoutes() []api.Route {
 	routes := []api.Route{
+		// plugin routes
 		api.Route{Method: "GET", Path: prefix + "/plugins", Handle: s.getPlugins},
 		api.Route{Method: "GET", Path: prefix + "/plugins/:type", Handle: s.getPlugins},
 		api.Route{Method: "GET", Path: prefix + "/plugins/:type/:name", Handle: s.getPlugins},
